finalize the ort c abi wrapper file name
