+++ base/string_utils.h
@@ -3,7 +3,7 @@
 #pragma once
 #include <sstream>
 #include <vector>
-#include "onnxruntime_cpp_api_legacy.h"
+#include "ort_c_to_cpp.h"
 
 template <typename T>
 inline void MakeStringInternal(std::ostringstream& ss, const T& t) noexcept {
